+++ docs/_toc.yml
@@ -42,6 +42,7 @@ parts:
           - file: notebooks/sax_02_model_extraction
           - file: notebooks/sax_03_variability_analysis
           - file: notebooks/lumerical_2_interconnect
+          - file: notebooks/vlsir_netlist
       - file: notebooks/20_schematic_driven_layout
   - caption: Verification
     chapters:
